darwin: fix build issue
This commit fixes issues building darwin binaries (darwin/amd64,arm64).  It
also adds unit tests and lint test on CI.

cgo should be enabled for darwin build because dependent package
(moby/sys/mountinfo) failed to lookup mounted volume, thus nerdctl fails
to unmount volumes.
