+++ .github/workflows/release.yml
@@ -7,8 +7,29 @@ on:
 env:
   GO111MODULE: on
 jobs:
+  osx-release:
+    runs-on: macos-11
+    timeout-minutes: 20
+    steps:
+    - uses: actions/setup-go@v2
+      with:
+        go-version: 1.17.x
+    - uses: actions/checkout@v2
+    - name: "install packages (osx)"
+      if: runner.os == 'macOS'
+      run: |
+        brew install gnu-tar
+        cp /usr/local/bin/gtar /usr/local/bin/tar
+    - name: "Compile binaries"
+      run: make artifacts
+    - name: upload artifact
+      uses: actions/upload-artifact@v2
+      with:
+        name: osx-artifacts
+        path: ${{ github.workspace }}/${{env.working-directory}}/_output/nerdctl-*
   release:
     runs-on: ubuntu-20.04
+    needs: osx-release
     timeout-minutes: 20
     steps:
     - uses: actions/setup-go@v2
@@ -17,6 +38,16 @@ jobs:
     - uses: actions/checkout@v2
     - name: "Compile binaries"
       run: make artifacts
+    - name: "downloading artifacts from external jobs"
+      uses: actions/download-artifact@v2
+      with:
+        name: osx-artifacts
+        path: osx-artifacts
+    - name: "place and display artifacts"
+      run: |
+        ls -lR osx-artifacts
+        mv osx-artifacts/nerdctl-* _output/
+        ls -lR _output
     - name: "SHA256SUMS"
       run: |
         ( cd _output; sha256sum nerdctl-* ) | tee /tmp/SHA256SUMS
