+++ .github/workflows/test.yml
@@ -26,8 +26,11 @@ jobs:
         go-mod-tidy-flags: -compat=1.17
 
   golangci-lint:
-    runs-on: ubuntu-20.04
+    runs-on: ${{ matrix.os }}
     timeout-minutes: 20
+    strategy:
+      matrix:
+        os: [ubuntu-20.04, macos-10.15]
     steps:
     - uses: actions/checkout@v2
       with:
@@ -39,8 +42,11 @@ jobs:
         args: --verbose
 
   test-unit:
-    runs-on: ubuntu-20.04
+    runs-on: ${{ matrix.os }}
     timeout-minutes: 20
+    strategy:
+      matrix:
+        os: [ubuntu-20.04, macos-10.15]
     steps:
     - uses: actions/setup-go@v2
       with:
@@ -100,7 +106,10 @@ jobs:
       run: docker run -t --rm --privileged test-integration-rootless-port-slirp4netns
 
   cross:
-    runs-on: ubuntu-20.04
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-20.04, macos-11]
     timeout-minutes: 30
     steps:
     - uses: actions/setup-go@v2
@@ -109,6 +118,11 @@ jobs:
     - uses: actions/checkout@v2
       with:
         fetch-depth: 1
+    - name: "install packages (osx)"
+      if: runner.os == 'macOS'
+      run: |
+        brew install gnu-tar
+        cp /usr/local/bin/gtar /usr/local/bin/tar
     - name: "Cross"
       run: make artifacts
