+++ cmd/nerdctl/run.go
@@ -24,7 +24,6 @@ import (
 	"fmt"
 	"net/url"
 	"os"
-	"path"
 	"path/filepath"
 	"runtime"
 	"strings"
@@ -698,36 +697,6 @@ func generateRootfsOpts(ctx context.Context, client *containerd.Client, platform
 		opts = append(opts, oci.WithRootFSReadonly())
 	}
 	return opts, cOpts, ensured, nil
-}
-
-// withBindMountHostProcfs replaces procfs mount with rbind.
-// Required for --pid=host on rootless.
-//
-// https://github.com/moby/moby/pull/41893/files
-// https://github.com/containers/podman/blob/v3.0.0-rc1/pkg/specgen/generate/oci.go#L248-L257
-func withBindMountHostProcfs(_ context.Context, _ oci.Client, _ *containers.Container, s *oci.Spec) error {
-	for i, m := range s.Mounts {
-		if path.Clean(m.Destination) == "/proc" {
-			newM := specs.Mount{
-				Destination: "/proc",
-				Type:        "bind",
-				Source:      "/proc",
-				Options:     []string{"rbind", "nosuid", "noexec", "nodev"},
-			}
-			s.Mounts[i] = newM
-		}
-	}
-
-	// Remove ReadonlyPaths for /proc/*
-	newROP := s.Linux.ReadonlyPaths[:0]
-	for _, x := range s.Linux.ReadonlyPaths {
-		x = path.Clean(x)
-		if !strings.HasPrefix(x, "/proc/") {
-			newROP = append(newROP, x)
-		}
-	}
-	s.Linux.ReadonlyPaths = newROP
-	return nil
 }
 
 func withCustomResolvConf(src string) func(context.Context, oci.Client, *containers.Container, *oci.Spec) error {
