+++ cmd/nerdctl/run_cgroup_windows.go
@@ -1,3 +1,6 @@
+//go:build windows
+// +build windows
+
 /*
    Copyright The containerd Authors.
