+++ cmd/nerdctl/run_gpus.go
@@ -1,3 +1,6 @@
+//go:build !darwin
+// +build !darwin
+
 /*
    Copyright The containerd Authors.
