+++ go.mod
@@ -165,3 +165,5 @@ require (
 
 // Temporary fork for avoiding importing patent-protected code: https://github.com/hashicorp/golang-lru/issues/73
 replace github.com/hashicorp/golang-lru => github.com/ktock/golang-lru v0.5.5-0.20211029085301-ec551be6f75c
+
+replace github.com/containerd/containerd => github.com/ukontainer/containerd v1.5.1-0.20220121000121-190b8b350994
