+++ pkg/netutil/cni_plugin_darwin.go
@@ -1,6 +1,3 @@
-//go:build freebsd || linux || darwin
-// +build freebsd linux darwin
-
 /*
    Copyright The containerd Authors.
 
@@ -17,32 +14,12 @@
    limitations under the License.
 */
 
-package main
+package netutil
 
-import (
-	"fmt"
-	"os"
-	"syscall"
+// TODO: empty for darwin
+type darwinIpamConfig struct {
+}
 
-	"golang.org/x/term"
-)
-
-func readPassword() (string, error) {
-	var fd int
-	if term.IsTerminal(syscall.Stdin) {
-		fd = syscall.Stdin
-	} else {
-		tty, err := os.Open("/dev/tty")
-		if err != nil {
-			return "", fmt.Errorf("error allocating terminal: %w", err)
-		}
-		defer tty.Close()
-		fd = int(tty.Fd())
-	}
-	bytePassword, err := term.ReadPassword(fd)
-	if err != nil {
-		return "", fmt.Errorf("error reading password: %w", err)
-	}
-
-	return string(bytePassword), nil
+func newDarwinIPAMConfig() *darwinIpamConfig {
+	return &darwinIpamConfig{}
 }
