+++ pkg/statsutil/stats_linux.go
@@ -167,3 +167,12 @@ func calculateCgroupNetwork(links []netlink.Link) (float64, float64) {
 	}
 	return rx, tx
 }
+
+func calculateMemPercent(limit float64, usedNo float64) float64 {
+	// Limit will never be 0 unless the container is not running and we haven't
+	// got any data from cgroup
+	if limit != 0 {
+		return usedNo / limit * 100.0
+	}
+	return 0
+}
